tools/scribe/actions/roadmap_populator_action.py
import json
import yaml
from pathlib import Path
from .base import BaseAction
from tools.scribe.core.atomic_write import atomic_write
from datetime import datetime
from typing import Dict, Any

class RoadmapPopulatorAction(BaseAction):
    def execute(self, file_content: str, match, file_path: str, params: Dict[str, Any]) -> str:
        # Parse JSON-LD or initialize from template
        jsonld_path = Path(file_path).with_suffix('.jsonld')
        if not jsonld_path.exists():
            # Load v2 template and initialize
            template = self._load_template('roadmap-template-v2.jsonld')
            data = self._populate_initial_data(template, params)
        else:
            with open(jsonld_path, 'r') as f:
                data = json.load(f)
        
        # Add/update timestamps
        timestamp = datetime.now().isoformat()
        data['date-modified'] = timestamp
        if 'date-created' not in data:
            data['date-created'] = timestamp
        
        # Validate structure
        if not self._validate_jsonld(data):
            # Fallback to minimal valid
            data = self._minimal_valid_jsonld(params)
        
        # Save JSON-LD atomically
        atomic_write(jsonld_path, json.dumps(data, indent=2))
        
        # Generate MD twin
        md_content = self._generate_md_twin(data)
        md_path = jsonld_path.with_suffix('.md')
        atomic_write(md_path, md_content)
        
        return file_content  # Original unchanged

    def _load_template(self, template_name: str) -> Dict:
        # Load JSON-LD template
        pass

    def _populate_initial_data(self, template: Dict, params: Dict) -> Dict:
        # Populate with params (e.g., project name, phases)
        pass

    def _validate_jsonld(self, data: Dict) -> bool:
        # Basic SHACL-like validation
        pass

    def _minimal_valid_jsonld(self, params: Dict) -> Dict:
        # Safe fallback structure
        return {'@context': '...', 'phases': []}

    def _generate_md_twin(self, data: Dict) -> str:
        # Convert to styled MD (trees, bars)
        pass 